requirements.txt
fastapi
uvicorn
streamlit
pandas
numpy
google-cloud-bigquery
google-cloud-storage
google-cloud-pubsub
google-cloud-aiplatform
python-dotenv
pydantic
plotly
